Created base module style to group all similiar logic
